Invoice calculating algorithm implementation. 17 passed test cases for the algorithm.

+++ schoolaccountingstudio-instruments/src/com/artiomastashonak/schoolaccountingstudio/invoice/Item.java
@@ -6,5 +6,5 @@ public record Item(String measureUnit,
                    String description,
                    short vat,
                    double price,
-                   float discount1,
-                   float discount2) { }
+                   double discount1,
+                   double discount2) { }
